[Bluetooth] Add adapter.setChangeListener API
This API permits the application to be notified of properties that are
changed in the adapter, for now, the name, the visibility and the powered
state are defined.

As these properties were already kept track of, it was just a matter of
caling the correct callbacks when they some of them change.