[Bluetooth] Ignore the error if the adapter is already enabled
In the tizen case, when the adapter is already up, but we didn't
get all the objects yet and we receive a setPowered() call, we may
try to enable an already enabled adapter, and it will fail.

We convert the error to an POSIX error so the dependence on tizen
headers are contained.